🐛 Fix: sample 이미지 동작 수정
