+++ public/assets/images/crew-sample/index.tsx
@@ -3,5 +3,10 @@ import ImgCrewSample02 from './crew-sample-2.jpg';
 import ImgCrewSample03 from './crew-sample-3.jpg';
 
 const ImgCrewSamples = [ImgCrewSample01, ImgCrewSample02, ImgCrewSample03];
+const ImgCrewSampleUrls = [
+  'https://crewcrew.s3.ap-northeast-2.amazonaws.com/crew/0e05d971-15a8-4a32-bf03-80d12cae392e',
+  'https://crewcrew.s3.ap-northeast-2.amazonaws.com/crew/eb35c35d-829a-402b-8019-29e42f91589f',
+  'https://crewcrew.s3.ap-northeast-2.amazonaws.com/crew/471b3b3b-b23c-48e8-8e6b-9a7ec31e1917',
+];
 
-export default ImgCrewSamples;
+export default ImgCrewSampleUrls;
