+++ src/components/common/input/file-input-wrap/index.tsx
@@ -7,7 +7,7 @@ import FileSample from './file-sample';
 
 export interface FileInputProps {
   value: File | string | null;
-  sample: StaticImageData[];
+  sample: string[];
   onChange: (newValue: File | string | null) => void;
 }
 
@@ -32,21 +32,9 @@ export default function FileInputWrap({ value, sample, onChange }: FileInputProp
 
   return (
     <div className="container flex max-w-[1200px] gap-4">
-      <FileSample
-        image="https://crewcrew.s3.ap-northeast-2.amazonaws.com/crew/0e05d971-15a8-4a32-bf03-80d12cae392e"
-        onChange={handleChange}
-        isBlur={isOtherSelected}
-      />
-      <FileSample
-        image="https://crewcrew.s3.ap-northeast-2.amazonaws.com/crew/eb35c35d-829a-402b-8019-29e42f91589f"
-        onChange={handleChange}
-        isBlur={isOtherSelected}
-      />
-      <FileSample
-        image="https://crewcrew.s3.ap-northeast-2.amazonaws.com/crew/471b3b3b-b23c-48e8-8e6b-9a7ec31e1917"
-        onChange={handleChange}
-        isBlur={isOtherSelected}
-      />
+      <FileSample image={sample[0]} onChange={handleChange} isBlur={isOtherSelected} />
+      <FileSample image={sample[1]} onChange={handleChange} isBlur={isOtherSelected} />
+      <FileSample image={sample[2]} onChange={handleChange} isBlur={isOtherSelected} />
       <FileInput value={fileValue} onChange={handleFileInput} isBlur={isSampleSelected} />
     </div>
   );
